refactor: 레이어간의 의존성을 줄이기 위해 패키지 구조 변경
