+++ src/main/java/com/market/application/service/LoginService.java
@@ -1,0 +1,8 @@
+package com.market.application.service;
+
+import com.market.application.dto.LoginResponseDto;
+import com.market.application.dto.SignInRequestDto;
+
+public interface LoginService {
+	LoginResponseDto login(SignInRequestDto dto);
+}
