+++ src/main/java/com/market/global/error/ErrorController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.market.auth.exception.UnauthorizedException;
+import com.market.global.exception.UnauthorizedException;
 
 @RestControllerAdvice
 public class ErrorController {
