+++ src/main/java/com/market/global/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.market.auth.exception;
+package com.market.global.exception;
 
 public class UnauthorizedException extends RuntimeException {
 	private static final String message = "인증되지 않은 사용자입니다.";
