+++ src/main/java/com/market/repository/mapper/RedisTemplateMapper.java
@@ -1,15 +1,17 @@
-package com.market.auth.repository;
+package com.market.repository.mapper;
 
 import java.util.Optional;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
 
+import com.market.repository.interfaces.RedisRepository;
+
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Repository
-public class RedisTemplateRepository implements RedisRepository {
+public class RedisTemplateMapper implements RedisRepository {
 	private final RedisTemplate<String, String> redisTemplate;
 
 	@Override
