--- src/main/java/com/market/user/service/LoginService.java
@@ -1,8 +1,0 @@
-package com.market.user.service;
-
-import com.market.user.controller.LoginResponse;
-import com.market.user.controller.dto.SignInRequestDto;
-
-public interface LoginService {
-	LoginResponse login(SignInRequestDto dto);
-}
