+++ src/main/resources/mybatis/mapper/UserMapper.xml
@@ -1,21 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.market.user.repository.UserRepository">
-    <insert id="insertUser" parameterType="User">
-        INSERT INTO USER (email, password, name, tel)
-        VALUES (#{email}, #{password}, #{name}, #{tel})
-    </insert>
-    <select id="findByEmail" resultType="User">
-        SELECT id, email, password, name, tel
-        FROM USER
-        WHERE email = #{email}
-    </select>
-    <select id="findByEmailAndPassword" resultType="User">
-        SELECT id, email, password, name, tel
-        FROM USER
-        WHERE email = #{email}
-          AND password = #{password}
-    </select>
+<mapper namespace="com.market.repository.mapper.UserMapper">
+	<insert id="insertUser" parameterType="User">
+		INSERT INTO USER (email, password, name, tel)
+		VALUES (#{email}, #{password}, #{name}, #{tel})
+	</insert>
+	<select id="findByEmail" resultType="User">
+		SELECT id, email, password, name, tel
+		FROM USER
+		WHERE email = #{email}
+	</select>
+	<select id="findByEmailAndPassword" resultType="User">
+		SELECT id, email, password, name, tel
+		FROM USER
+		WHERE email = #{email}
+		  AND password = #{password}
+	</select>
 </mapper>
